refactor: split cli command/register to fix and make testing easier
