+++ src/cli/commands/deploy/index.ts
@@ -1,0 +1,2 @@
+export * from "./deploy";
+export { default as registerDeploy } from "./register";
