+++ src/cli/commands/init/index.ts
@@ -1,0 +1,2 @@
+export * from "./init";
+export { default as registerInit } from "./register";
