+++ src/cli/commands/init/init.spec.ts
@@ -1,9 +1,9 @@
 import fs from "fs";
 import mockFs from "mock-fs";
 import { init } from "./init";
-import { DEFAULT_CONFIG } from "../../config";
-import { swaCliConfigFilename } from "../../core/utils";
-import { convertToNativePaths, convertToUnixPaths } from "../../jest.helpers.";
+import { DEFAULT_CONFIG } from "../../../config";
+import { swaCliConfigFilename } from "../../../core/utils";
+import { convertToNativePaths, convertToUnixPaths } from "../../../jest.helpers.";
 
 jest.mock("prompts", () => jest.fn());
