+++ src/cli/commands/login/index.ts
@@ -1,0 +1,2 @@
+export * from "./login";
+export { default as registerLogin, addSharedLoginOptionsToCommand } from "./register";
