+++ src/cli/commands/start/index.ts
@@ -1,0 +1,2 @@
+export * from "./start";
+export { default as registerStart } from "./register";
